+++ src/main/java/insane96mcp/survivalreimagined/module/experience/feature/PlayerExperience.java
@@ -1,11 +1,14 @@
 package insane96mcp.survivalreimagined.module.experience.feature;
 
+import com.ezylang.evalex.Expression;
+import com.ezylang.evalex.data.EvaluationValue;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.survivalreimagined.module.Modules;
+import insane96mcp.survivalreimagined.utils.LogHelper;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.GameRules;
 import net.minecraftforge.event.TickEvent;
@@ -14,13 +17,13 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 @Label(name = "Player Experience", description = "Changes the experience lost on death and xp per level required.")
 @LoadFeature(module = Modules.Ids.EXPERIENCE)
 public class PlayerExperience extends Feature {
-	//TODO Change this into a formula
-	@Config(min = -1)
-	@Label(name = "Flat XP to next level", description = """
-						The experience required to level up will be fixed to this value.
-						Set to 0 to disable.
+
+	@Config
+	@Label(name = "Level Scaling Formula", description = """
+						The experience required to level up is calculated by this formula. The variable available is 'current_level'
+						Set to empty to disable.
 						Incompatible with Allurement's 'Remove level Scaling' or any other mod that does something similar""")
-	public static Integer flatLevelScaling = 50;
+	public static String levelScalingFormula = "3 * (current_level + 1)";
 	@Config(min = -1d, max = 1d)
 	@Label(name = "Dropped Experience on Death", description = """
 						On death, players will drop this percentage of experience instead of max 7 levels. Setting to -1 will disable this.
@@ -48,12 +51,23 @@ public class PlayerExperience extends Feature {
 	/**
 	 * Returns -1 when the module/feature is not enabled, otherwise the calculated experience required to level up
 	 */
-	public static int getBetterScalingLevel(int experienceLevel) {
+	public static int getBetterScalingLevel(int currentLevel) {
 		if (!isEnabled(PlayerExperience.class)
-				|| flatLevelScaling == 0)
+				|| levelScalingFormula.isEmpty())
 			return -1;
 
-		return flatLevelScaling;
+		Expression expression = new Expression(levelScalingFormula);
+		try {
+			//noinspection ConstantConditions
+			EvaluationValue result = expression
+					.with("current_level", currentLevel)
+					.evaluate();
+			return result.getNumberValue().intValue();
+		}
+		catch (Exception ex) {
+			LogHelper.error("Failed to evaluate or parse eating speed formula: %s", expression);
+			return vanillaXpNeededForNextLevel(currentLevel);
+		}
 	}
 
 	//Instead of using experienceTotal, calculate the xp from the xp bar and level since experienceTotal doesn't get updated on level consume
@@ -76,7 +90,10 @@ public class PlayerExperience extends Feature {
 		int betterScalingXp = getBetterScalingLevel(currentLevel);
 		if (betterScalingXp != -1)
 			return betterScalingXp;
+		return vanillaXpNeededForNextLevel(currentLevel);
+	}
 
+	private static int vanillaXpNeededForNextLevel(int currentLevel) {
 		if (currentLevel >= 30) {
 			return 112 + (currentLevel - 30) * 9;
 		} else {